app/dticlustering/lib/requirements.txt
pyyaml
tqdm
pandas
Pillow<10.0
scipy
kornia
scikit-learn
torch==1.12.1+cu116
torchvision==0.13.1+cu116
torchaudio==0.12.1
matplotlib
seaborn
--extra-index-url https://download.pytorch.org/whl/cu116
visdom
django-environ